Se ha implementado el chequeo del goal de calorias diarias y la asignacion

+++ src/main/java/com/fitnessapp/fitapp_api/gamification/service/PointsCalculationService.java
@@ -4,5 +4,5 @@ import com.fitnessapp.fitapp_api.gamification.dto.PCActivityRequestDTO;
 import com.fitnessapp.fitapp_api.routeexecution.model.RouteExecution;
 
 public interface PointsCalculationService {
-    double calculatePoints(PCActivityRequestDTO pcActivityRequestDTO);
+    long calculatePoints(PCActivityRequestDTO pcActivityRequestDTO);
 }
